Move layout package to cmp package
