+++ admin/App.js
@@ -7,7 +7,7 @@
 
 import {Component} from 'react';
 import {XH, HoistComponent} from 'hoist/core';
-import {frame} from 'hoist/layout';
+import {frame} from 'hoist/cmp/layout';
 import {navbar, navbarGroup, navbarHeading, button} from 'hoist/kit/blueprint';
 import {logoutButton, lockoutPanel, panel, tabContainer, themeToggleButton, refreshButton} from 'hoist/cmp';
 import {Icon} from 'hoist/icon';
