+++ admin/tabs/about/AboutPanel.js
@@ -7,7 +7,7 @@
 
 import React, {Component} from 'react';
 import {XH, HoistComponent} from 'hoist/core';
-import {div, h3, h4, table, tbody, tr, th, td} from 'hoist/layout';
+import {div, h3, h4, table, tbody, tr, th, td} from 'hoist/cmp/layout';
 
 import './AboutPanel.scss';
