+++ admin/tabs/configs/differ/ConfigDifferDetail.js
@@ -8,7 +8,7 @@ import {Component} from 'react';
 import {keys, toString} from 'lodash';
 import {HoistComponent, elemFactory} from 'hoist/core';
 import {button, dialog} from 'hoist/kit/blueprint';
-import {filler, table, tbody, tr, th, td} from 'hoist/layout';
+import {filler, table, tbody, tr, th, td} from 'hoist/cmp/layout';
 import {panel, toolbar} from 'hoist/cmp';
 import {Icon} from 'hoist/icon';
