+++ admin/tabs/ehcache/EhCachePanel.js
@@ -8,7 +8,7 @@ import {Component} from 'react';
 import {button} from 'hoist/kit/blueprint';
 import {HoistComponent} from 'hoist/core';
 import {grid} from 'hoist/cmp/grid';
-import {filler} from 'hoist/layout';
+import {filler} from 'hoist/cmp/layout';
 import {panel, refreshButton, storeCountLabel, storeFilterField, toolbar, toolbarSep} from 'hoist/cmp';
 import {Icon} from 'hoist/icon';
