+++ admin/tabs/logging/viewer/LogViewerToolbar.js
@@ -8,7 +8,7 @@ import {Component} from 'react';
 import {elemFactory, HoistComponent} from 'hoist/core';
 import {button} from 'hoist/kit/blueprint';
 import {checkField, label, numberField, textField, toolbar} from 'hoist/cmp';
-import {filler} from 'hoist/layout';
+import {filler} from 'hoist/cmp/layout';
 import {Icon} from 'hoist/icon';
 
 @HoistComponent()
