+++ admin/tabs/monitor/Tile.js
@@ -7,7 +7,7 @@
 
 import {Component} from 'react';
 import {HoistComponent, elemFactory} from 'hoist/core';
-import {vbox, div} from 'hoist/layout';
+import {vbox, div} from 'hoist/cmp/layout';
 import {getRelativeTimestamp} from 'hoist/cmp';
 import {pluralize} from 'hoist/utils/JsUtils';
 import {Icon} from 'hoist/icon';
