+++ admin/tabs/services/ServicePanel.js
@@ -7,7 +7,7 @@
 import {Component} from 'react';
 import {HoistComponent} from 'hoist/core';
 import {grid} from 'hoist/cmp/grid';
-import {filler} from 'hoist/layout';
+import {filler} from 'hoist/cmp/layout';
 import {button} from 'hoist/kit/blueprint';
 import {panel, refreshButton, storeCountLabel, storeFilterField, toolbar, toolbarSep} from 'hoist/cmp';
 import {Icon} from 'hoist/icon';
