+++ admin/tabs/users/UserPanel.js
@@ -7,7 +7,7 @@
 import {Component} from 'react';
 import {HoistComponent} from 'hoist/core';
 import {grid} from 'hoist/cmp/grid';
-import {filler} from 'hoist/layout';
+import {filler} from 'hoist/cmp/layout';
 import {panel, refreshButton, storeCountLabel, storeFilterField, toolbar} from 'hoist/cmp';
 
 import {UserModel} from './UserModel';
