+++ app/impl/LoginPanel.js
@@ -7,7 +7,7 @@
 
 import {Component} from 'react';
 import {XH, elemFactory, HoistComponent} from 'hoist/core';
-import {vspacer, box, filler, viewport} from 'hoist/layout';
+import {vspacer, box, filler, viewport} from 'hoist/cmp/layout';
 import {button, text} from 'hoist/kit/blueprint';
 import {panel, textField, toolbar} from 'hoist/cmp';
 import {observable, computed, setter} from 'hoist/mobx';
