+++ app/impl/UpdateBar.js
@@ -7,7 +7,7 @@
 import {Component} from 'react';
 import {XH, elemFactory, HoistComponent} from 'hoist/core';
 import {toolbar} from 'hoist/cmp';
-import {div} from 'hoist/layout';
+import {div} from 'hoist/cmp/layout';
 import {button} from 'hoist/kit/blueprint';
 import {Icon} from 'hoist/icon';
 import './UpdateBar.scss';
