+++ cmp/LockoutPanel.js
@@ -7,7 +7,7 @@
 
 import {Component} from 'react';
 import {HoistComponent, elemFactory} from 'hoist/core';
-import {box, viewport} from 'hoist/layout';
+import {box, viewport} from 'hoist/cmp/layout';
 import {PropTypes as PT} from 'prop-types';
 
 import './LockoutPanel.scss';
