+++ cmp/mask/Mask.js
@@ -7,7 +7,7 @@
 import {Component} from 'react';
 import {PropTypes as PT} from 'prop-types';
 import {HoistComponent, elemFactory} from 'hoist/core';
-import {box} from 'hoist/layout';
+import {box} from 'hoist/cmp/layout';
 import {Classes, overlay} from 'hoist/kit/blueprint';
 
 import './Mask.scss';
