+++ cmp/message/Message.js
@@ -6,7 +6,7 @@
  */
 import {Component} from 'react';
 import {HoistComponent, elemFactory} from 'hoist/core';
-import {filler} from 'hoist/layout';
+import {filler} from 'hoist/cmp/layout';
 import {toolbar} from 'hoist/cmp';
 import {dialog, dialogBody, button} from 'hoist/kit/blueprint';
