+++ cmp/panel/Panel.js
@@ -8,7 +8,7 @@ import {Component} from 'react';
 import {PropTypes as PT} from 'prop-types';
 import {castArray} from 'lodash';
 import {elemFactory, HoistComponent} from 'hoist/core';
-import {vframe, vbox} from 'hoist/layout';
+import {vframe, vbox} from 'hoist/cmp/layout';
 import {mask} from 'hoist/cmp';
 import {panelHeader} from './impl/PanelHeader';
