+++ cmp/panel/impl/PanelHeader.js
@@ -6,7 +6,7 @@
  */
 import {Component} from 'react';
 import {elemFactory, HoistComponent} from 'hoist/core';
-import {box, hbox} from 'hoist/layout';
+import {box, hbox} from 'hoist/cmp/layout';
 
 import './PanelHeader.scss';
