+++ cmp/resizable/Resizable.js
@@ -7,7 +7,7 @@
 import React, {Children, Component} from 'react';
 import {PropTypes as PT} from 'prop-types';
 import {elemFactory, HoistComponent} from 'hoist/core';
-import {box, hbox, vbox} from 'hoist/layout';
+import {box, hbox, vbox} from 'hoist/cmp/layout';
 
 import {ResizableModel} from './ResizableModel';
 import {dragger} from './impl/Dragger';
