+++ cmp/resizable/impl/Collapser.js
@@ -10,7 +10,7 @@ import {PropTypes as PT} from 'prop-types';
 import {elemFactory, HoistComponent} from 'hoist/core';
 import {Icon} from 'hoist/icon';
 import {button} from 'hoist/kit/blueprint';
-import {hbox, vbox} from 'hoist/layout';
+import {hbox, vbox} from 'hoist/cmp/layout';
 
 import './Collapser.scss';
