+++ cmp/tab/TabPane.js
@@ -7,7 +7,7 @@
 import {Component} from 'react';
 import {elem, elemFactory, HoistComponent} from 'hoist/core';
 import {Ref} from 'hoist/utils/Ref';
-import {frame} from 'hoist/layout';
+import {frame} from 'hoist/cmp/layout';
 
 /**
  * Wrapper for Components to be shown inside of a TabContainer tab. Provides the following:
