+++ cmp/toolbar/Toolbar.js
@@ -8,7 +8,7 @@ import {Component} from 'react';
 import {PropTypes as PT} from 'prop-types';
 import {button} from 'hoist/kit/blueprint';
 import {elemFactory, HoistComponent} from 'hoist/core';
-import {hbox, vbox} from 'hoist/layout';
+import {hbox, vbox} from 'hoist/cmp/layout';
 
 import './Toolbar.scss';
